+++ source/TestCommon/documents/release-notes/release-notes.md
@@ -2,7 +2,7 @@
 
 ## Version 4.2.0
 
-- Added property `DatabricksWarehouseSettings` to `IntegrationTestConfiguration` to support use of Databricks SQL Warehouse.
+- Added property `DatabricksSettings` to `IntegrationTestConfiguration` to support use of Databricks workspace and SQL Warehouse.
 
 ## Version 4.1.1